% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm.R
\name{ProjectManage}
\alias{ProjectManage}
\title{ProjectManage}
\usage{
ProjectManage()
}
\description{
A simple Project manage web server, designed for project manage.
In this web server,you can add,delete and edit the Project.
Yixuetongji Group developed this shiny App for his fans.
If you have any questions, pls feel free to contact me(dongliulou@126.com).
Furthermore, you can follow this guy by wechat:yixuetongji001.
}
\author{
Dming1024
}
