% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm.R
\name{ProjectManage}
\alias{ProjectManage}
\title{ProjectManage}
\usage{
ProjectManage()
}
\description{
Add new users into the database
in this web server,including add,delete and edit the Project.
Any rows with duplicated row names will be dropped with the first one being
kepted.
}
\author{
FanQiangqiang
}
